Move computeNeed from BURN utils to core/burn for shared use
