+++ src/core/burn.ts
@@ -164,3 +164,16 @@ export function calculatePlanetBurn(
 
   return burnValues;
 }
+
+export function computeNeed(mat: MaterialBurn, resupplyDays: number) {
+  const production = mat.dailyAmount;
+  const isInf = production >= 0;
+  const days = isInf ? 1000 : mat.daysLeft;
+  if (days > resupplyDays || production > 0) {
+    return 0;
+  }
+  const need = Math.ceil((days - resupplyDays) * production);
+  // This check is needed to prevent a "-0" value that can happen
+  // in situations like: 0 * -0.25 => -0.
+  return need === 0 ? 0 : need;
+}
