+++ src/features/XIT/BURN/utils.ts
@@ -1,4 +1,4 @@
-import { BurnValues, MaterialBurn, PlanetBurn } from '@src/core/burn';
+import { BurnValues, PlanetBurn } from '@src/core/burn';
 import { materialsStore } from '@src/infrastructure/prun-api/data/materials';
 import { sortMaterials } from '@src/core/sort-materials';
 
@@ -16,17 +16,4 @@ export function countDays(burn: BurnValues) {
     }
   }
   return days;
-}
-
-export function computeNeed(mat: MaterialBurn, resupplyDays: number) {
-  const production = mat.dailyAmount;
-  const isInf = production >= 0;
-  const days = isInf ? 1000 : mat.daysLeft;
-  if (days > resupplyDays || production > 0) {
-    return 0;
-  }
-  const need = Math.ceil((days - resupplyDays) * production);
-  // This check is needed to prevent a "-0" value that can happen
-  // in situations like: 0 * -0.25 => -0.
-  return need === 0 ? 0 : need;
 }
